+++ middleware/log.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"context"
 	"fmt"
-	"io"
 	mathRand "math/rand"
 	"net"
 	"net/http"
@@ -17,25 +16,18 @@ import (
 const logIDKey = string(log.CtxKey)
 
 // Log is a middleware that logs requests/responses.
-func Log(wrappedHandler http.HandlerFunc, domain string, logOutput io.Writer) http.HandlerFunc {
-	// We want different requests to share the same logger backed by the same circular buffere for storing logs.
-	// That way, if someone makes one request and it succeds then they make another one that errors.
-	// When the logs are been flushed for the request that errored, the logs for the request that succeeded will also be flushed.
-	// Thus app developers can be able to correlate issues/logs in much better way.
+func Log(wrappedHandler http.HandlerFunc, domain string, l log.Logger) http.HandlerFunc {
+	// We pass the logger as an argument so that the middleware can share the same logger as the app.
+	// That way, if the app logs an error, the middleware logs are also flushed.
+	// This makes debugging easier for developers.
 	//
 	// However, each request should get its own context. That's why we call `logger.WithCtx` for every request.
-	mLogger := log.New(
-		context.Background(),
-		logOutput,
-		// enought to hold messages for 5reqs/sec for 15minutes.
-		5*60*15,
-	)
 
 	mathRand.Seed(time.Now().UTC().UnixNano())
 
 	return func(w http.ResponseWriter, r *http.Request) {
 		ctx := r.Context()
-		logger := mLogger.WithCtx(ctx)
+		reqL := l.WithCtx(ctx)
 
 		{
 			// set cookie/headers/ctx for logID.
@@ -83,13 +75,13 @@ func Log(wrappedHandler http.HandlerFunc, domain string, logOutput io.Writer) ht
 				// Only log 10% of the errors.
 				shouldLog := mathRand.Intn(100) > 90
 				if shouldLog {
-					logger.Error(nil, flds)
+					reqL.Error(nil, flds)
 				}
 			} else if lrw.code >= http.StatusBadRequest {
 				// both client and server errors.
-				logger.Error(nil, flds)
+				reqL.Error(nil, flds)
 			} else {
-				logger.Info(flds)
+				reqL.Info(flds)
 			}
 		}()