+++ middleware/panic_test.go
@@ -2,7 +2,9 @@ package middleware
 
 import (
 	"bytes"
+	"context"
 	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -11,6 +13,7 @@ import (
 	"testing"
 
 	"github.com/komuw/ong/errors"
+	"github.com/komuw/ong/log"
 
 	"github.com/akshayjshah/attest"
 )
@@ -33,12 +36,16 @@ func handlerThatPanics(msg string, shouldPanic bool, err error) http.HandlerFunc
 func TestPanic(t *testing.T) {
 	t.Parallel()
 
+	getLogger := func(w io.Writer) log.Logger {
+		return log.New(context.Background(), w, 500)
+	}
+
 	t.Run("ok if no panic", func(t *testing.T) {
 		t.Parallel()
 
 		logOutput := &bytes.Buffer{}
 		msg := "hello"
-		wrappedHandler := Panic(handlerThatPanics(msg, false, nil), logOutput)
+		wrappedHandler := Panic(handlerThatPanics(msg, false, nil), getLogger(logOutput))
 
 		rec := httptest.NewRecorder()
 		req := httptest.NewRequest(http.MethodGet, "/someUri", nil)
@@ -55,7 +62,7 @@ func TestPanic(t *testing.T) {
 
 		logOutput := &bytes.Buffer{}
 		msg := "hello"
-		wrappedHandler := Panic(handlerThatPanics(msg, true, nil), logOutput)
+		wrappedHandler := Panic(handlerThatPanics(msg, true, nil), getLogger(logOutput))
 
 		rec := httptest.NewRecorder()
 		req := httptest.NewRequest(http.MethodGet, "/someUri", nil)
@@ -85,7 +92,7 @@ func TestPanic(t *testing.T) {
 		msg := "hello"
 		errMsg := "99 problems"
 		err := errors.New(errMsg)
-		wrappedHandler := Panic(handlerThatPanics(msg, false, err), logOutput)
+		wrappedHandler := Panic(handlerThatPanics(msg, false, err), getLogger(logOutput))
 
 		rec := httptest.NewRecorder()
 		req := httptest.NewRequest(http.MethodGet, "/someUri", nil)
@@ -118,7 +125,7 @@ func TestPanic(t *testing.T) {
 		err := errors.New(msg)
 		// for this concurrency test, we have to re-use the same wrappedHandler
 		// so that state is shared and thus we can see if there is any state which is not handled correctly.
-		wrappedHandler := Panic(handlerThatPanics(msg, false, err), logOutput)
+		wrappedHandler := Panic(handlerThatPanics(msg, false, err), getLogger(logOutput))
 
 		runhandler := func() {
 			rec := httptest.NewRecorder()